requirements.txt
requests
beautifulsoup4
lxml
tqdm
markdownify
sentence-transformers
faiss-cpu
numpy
